agent/requirements.txt
dagger-io==0.9.3
docker==7.0.0
pyyaml==6.0.1
requests==2.31.0